pyproject.toml
[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "aa-hr-centre"
dynamic = ["version"]
authors = [
    {name = "Matteo Ghia", email = "matteo.ghia@yahoo.it"},
]
classifiers = [
    "Environment :: Web Environment",
	"Framework :: Django",
	"Framework :: Django :: 4",
	"Framework :: Django :: 4.2",
	"Operating System :: POSIX :: Linux",
	"Development Status :: 3 - Alpha",
	"Intended Audience :: Developers",
	"License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
	"Natural Language :: English",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.10",
	"Programming Language :: Python :: 3.11",
	"Programming Language :: Python :: 3.12",
	"Topic :: Internet :: WWW/HTTP",
	"Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
description = "HR Centre for Alliance Auth"
license = {file = "LICENSE"}
readme = "README.md"
keywords = [
    "allianceauth",
    "eveonline",
	"allianceauth_hr_centre",
    "hr",
	"hr_centre",
]
requires-python = "~=3.10"
dependencies = [
    "allianceauth~=4.6",
	"allianceauth-corptools~=2.9",
]

[project.urls]
Homepage = "https://github.com/Maestro-Zacht/aa-hr-centre"
Source = "https://github.com/Maestro-Zacht/aa-hr-centre"
Tracker = "https://github.com/Maestro-Zacht/aa-hr-centre/issues"
Changelog = "https://github.com/Maestro-Zacht/aa-hr-centre/releases"

[tool.flit.module]
name = "hrcentre"
